커맨드 라인에서 바우처 생성 함수 통합 및 바우처 도메인에서 입력값 검증

+++ src/main/java/prgms/spring_week1/domain/voucher/service/VoucherService.java
@@ -20,7 +20,7 @@ public class VoucherService {
         voucherRepository.insert(new FixedAmountVoucher(discountAmount));
     }
 
-    public void insertPercentDiscountVoucher(int discountPercent) {
+    public void insertPercentDiscountVoucher(Long discountPercent) {
         voucherRepository.insert(new PercentDiscountVoucher(discountPercent));
     }
